feat(peers): use a common way to bootstrap into the network for all the bins
