+++ sn_cli/Cargo.toml
@@ -22,7 +22,7 @@ harness = false
 default = ["metrics"]
 local-discovery=["sn_client/local-discovery"]
 metrics = ["sn_logging/process-metrics"]
-network-contacts = []
+network-contacts = ["sn_peers_acquisition/network-contacts"]
 open-metrics = ["sn_client/open-metrics"]
 
 [dependencies]
